hanghoadiaglog

+++ QuanLyQuanCafe/QuanLy/HangHoa.cs
@@ -8,6 +8,7 @@ using APP;
 using BUS;
 using DTO;
 using MaterialSkin.Controls;
+using QuanLyQuanCafe.Dialog;
 
 namespace QuanLyQuanCafe.QuanLy
 {
@@ -108,8 +109,11 @@ namespace QuanLyQuanCafe.QuanLy
 
             using (HangHoaBUS bus = new HangHoaBUS())
             {
-                string tenhanghoa = dataGridView1.CurrentRow.Cells[0].Value.ToString();
-                bus.DeleteHangHoa(tenhanghoa);
+                if (new XacNhan { Text = @"Bạn có chắc chắn muốn xóa hàng hóa " + dataGridView1.CurrentRow.Cells[0].Value }.ShowDialog() == DialogResult.Yes)
+                {
+                    string tenhanghoa = dataGridView1.CurrentRow.Cells[0].Value.ToString();
+                    bus.DeleteHangHoa(tenhanghoa);
+                }
             }
 
             RefreshHangHoa();
